updates to Detroix
removed a few lines of Code in the CSS and added more in both the HTML and CSS. Added images as well and trying to work on the bottom part of the website. Currently a mess but will update tomorrow. Make sure to check your code for the "closer blog" section to fix it.